+++ src/interface.tsx
@@ -1,8 +1,5 @@
 import * as React from 'react';
-import {
-  SelectCallback,
-  Sizes,
-} from 'react-bootstrap';
+import { SelectCallback, Sizes } from 'react-bootstrap';
 import { Moment } from 'moment';
 import CSS from 'csstype';
 
@@ -51,6 +48,7 @@ export interface IconProps {
   className?: string;
   onClick?: React.MouseEventHandler;
   style?: object;
+  cursor?: boolean;
 }
 
 export interface TooltipProps {
@@ -90,6 +88,7 @@ export interface UsageBarProps {
   hideRight?: boolean;
   showZeroMax?: boolean;
   withUnavailable?: boolean;
+  formatType?: string;
 }
 
 export interface LoaderProps {
@@ -106,6 +105,10 @@ export interface TabsProps {
   onSelect?: SelectCallback;
   activeKey?: string;
   id?: string;
+  limitNum?: number;
+  unmountOnExit?: boolean;
+  animation?: boolean;
+  mountOnEnter?: boolean;
 }
 
 export interface StepsProps {
@@ -306,11 +309,7 @@ export interface MultiVirtualSelectItem {
   label: string;
   value: number | string;
 }
-export type VirtualItem =
-  | MultiVirtualSelectItem
-  | object
-  | string
-  | number;
+export type VirtualItem = MultiVirtualSelectItem | object | string | number;
 export interface SelectCheckItemProps {
   onSelect: Function;
   selected: boolean;